Drop "vim-ruby" to prevent ctag execution errors in vim :wave:
Executing the ctag shortcut keys causes errors.
It seems that "vim-ruby" has benn causing the errors, unfortunately the plugin development seems not so active anymore.
I want to use the shortcut keys of ctag, so I think it is the timing to drop it.

The error was:
```
E117: Unknown function: RubyCursorTag
E116: Invalid arguments for function substitute
E15: Invalid expression: substitute(RubyCursorTag(),'^$',"\022\027",'')
```