Use stable "pyenv"
If "Python" is installed with `HEAD` version, the commit hash will be written in "pip" script.
Upgrading "Pyenv" will cause the commit hash to shift, failing to update "pip" and I have to reinstall all "Python" versions.
It is troublesome to handle. To that much there is no reason to me to use the `HEAD` version.

Error was:
```
/Users/machupicchubeta/.pyenv/shims/pip: line 21: /usr/local/Cellar/pyenv/HEAD-ec9fb54/libexec/pyenv: No such file or directory
/Users/machupicchubeta/.pyenv/shims/pip2: line 21: /usr/local/Cellar/pyenv/HEAD-ec9fb54/libexec/pyenv: No such file or directory
/Users/machupicchubeta/.pyenv/shims/pip3: line 21: /usr/local/Cellar/pyenv/HEAD-ec9fb54/libexec/pyenv: No such file or directory
```