Install `sleuthkit` with Homebrew explicitly
It seems to some command depends on `sleuthkit`.
Since it is cumbersome to find out which commands are depends on it, try installing `sleuthkit` explicitly to fix the error.

Error was:
```
sleuthkit: Java is required to install this formula.
Install AdoptOpenJDK with Homebrew Cask:
  brew cask install adoptopenjdk
Error: An unsatisfied requirement failed this build.
```