Make sure always to install "Command Line Tools" at every run setup script
When running `xcode-select --install` command, it may determine that the installation is done even if the SDK is missing.
In particular, after upgrading macOS, the SDK and other command directories often disappear, and even in that state, the installation was determined as completed.
Therefore, I decided to remove its whole directory and reinstall it every time I run the setup script.