Switch the shell prompt tool using from "Spaceship ZSH" to "Starship" :rocket:
I was satisfied with "Spaceship ZSH", but when I tried "Starship", it seemed to display faster, a bit maybe.
"Starship" allows me to set the timeout threshold for its executing commands, so -- for example -- it will be useful for git operations on directories cloned from repositories with a long history.
In such a case, "Spaceship ZSH" was annoying me which would wait for several seconds (sometime several tens of seconds) before displaying the prompt.

Refs.
- https://starship.rs/
- https://starship.rs/config/
- https://github.com/starship/starship
- https://github.com/spaceship-prompt/spaceship-prompt