Avoid difficulty in Typing with "VSCode Neovim"
When I use the "Indent Guides" Vim plugin with "VSCode Neovim", the text on the line who am typing is dark and difficult to see.
I use the "Indent-Rainbow" extension on "Visual Studio Code" to visualize indentation, so I don't need to load the Vim plugin in that editor environment.

Refs.
- https://github.com/nathanaelkane/vim-indent-guides
- https://marketplace.visualstudio.com/items?itemName=asvetliakov.vscode-neovim
- https://marketplace.visualstudio.com/items?itemName=oderwat.indent-rainbow