Use "sheldon", instead of "zplug"
When I start Terminal.app, I get an error message that "lib/termsupport" could not be loaded.

> [zplug] These repos have failed to load:
> - lib/termsupport

The same error is reported in an Issue of "zplug", which was reported in June 2020, but has not yet been commented on.

- https://github.com/zplug/zplug/issues/550

"zplug" has not been maintained for years. The latest git commit was in February 2020, and there have been no commits for almost two years.
The author says it no longer has the motivation to maintain "zplug" and is developing other CLI tool similar to.

- https://github.com/zplug/zplug/issues/537
- https://github.com/zplug/zplug/issues/579
- https://babarot.me/afx/

At first I thought about switching to that tool, but as I searched the web for alternative tools, "sheldon" seemed better to me.
The loading time seemed fast, and one of the deciding factors was that it could be installed using Homebrew.

- https://www.libhunt.com/r/zplug
- https://github.com/rossmacarthur/zsh-plugin-manager-benchmark#results
- https://formulae.brew.sh/formula/sheldon
- https://github.com/rossmacarthur/sheldon
- https://sheldon.cli.rs/Introduction.html
- https://ktrysmt.github.io/blog/migrate-zinit-to-sheldon/