Get ready to setup Mac computers with Apple silicon
Enable to setup not only Mac computers with an Intel processor but also Mac computers with an Apple silicon.

I referenced these web pages.
- https://zenn.dev/sprout2000/articles/aad599d3625242
- https://zenn.dev/ress/articles/069baf1c305523dfca3d
- https://oknm.jp/posts/e8cf774c2703c3677f0956ee10dc9ea2a8a9b47d
- https://titech-os.github.io/xv6-mac.html
- https://osxdaily.com/2020/12/04/how-install-rosetta-2-apple-silicon-mac/
- https://hkob.hatenablog.com/entry/2020/11/27/090000
- https://zenn.dev/hinastory/articles/71983c4ac8aa2d
- https://support.apple.com/en-us/HT211814
- https://forum.latenightsw.com/t/possible-for-a-script-to-test-whether-rosetta-2-is-installed/3207/6
- https://community.jamf.com/t5/jamf-pro/how-to-check-if-rosetta2-is-installed-macos-11-5/m-p/242580
- https://stackoverflow.com/questions/64908116/correct-archflags-value-on-apple-silicon
- https://gitlab.kitware.com/cmake/cmake/-/issues/20989
- https://news.ycombinator.com/item?id=25183120

I don't know why, but with Rosetta 2 installed, I get an error when I try to switch between arm64 and x86_64 arch with the `exec arch -arch` command as in the web pages I referred to.
It would be useful to be able to switch between arm64 and x86_64 arch with the command In the future, I will find a snippet or something to improve this so that it can be used without error.

Error was:
```
$ uname -m

arm64

$ exec arch -arch x86_64 "$SHELL"

arch: posix_spawnp: /opt/homebrew/bin/zsh: Bad CPU type in executable

[Process completed]
```