Make sure to initialize programming environment tools at shell startup
I remember that when I first started using "sheldon", the Programming environment tools were initialized at shell startup. Umm ....
But I recently noticed that they were not initialized and not added to the `PATH` environment variable, so that only the commands installed by "Homebrew" could be found.
This was very inconvenient.

This commit will ensure that the programming environment tools are initialized via "sheldon".
It will now switch to the programming language version specified in the `.xxx-version` file.