Stop using Oh My Zsh Docker plugin via sheldon
It seems that the use case of using only the Oh My Zsh plugin is not intended, and an error occurs when loading the plugin in sheldon.
I changed the directory arrangement in `SHELDON_CONFIG_DIR` and `SHELDON_DATA_DIR`, and now the error occurs when loading the plugin.

The error was:
```
/Users/machupicchubeta/.local/share/sheldon/zsh/repos/github.com/ohmyzsh/ohmyzsh/plugins/docker/docker.plugin.zsh:60: no such file or directory: /completions/_docker
```