Move GOENV_ROOT and GOENV_GOPATH_PREFIX under XDG Base Directory
Refs.
- https://github.com/go-nv/goenv/blob/56c9dec4e8184f16b4af886134b44b9eb2edf9b7/ENVIRONMENT_VARIABLES.md

In addition, I added `GOENV_ROOT/shims` to the PATH environment variable to give preference to Go installed with goenv over Homebrew.
For example, rbenv, nodenv, and pyenv add their respective `shims` directories to the top of the PATH environment variable, but goenv did not, and only Go installed with Homebrew was found.