Create a new configuration file for Neovim in the lua language
So far, Vim's configuration file was used as it is in Neovim.
However, I was concerned that Neovim takes a long time to start up, so I made a major revision and created a new configuration file.
In that case, some items were used from Vim's configuration file, but plugins were selected again.

Refs.
- https://neovim.io/doc/user/index.html
- https://neovim.io/doc/user/lua-guide.html#lua-guide
- https://neovim.io/doc/user/lua.html#lua

I refer to these articles.
- https://speakerdeck.com/joker1007/neovimdezuo-ruzui-xin-rubykai-fa-huan-jing-2023?slide=3
- https://joker1007.hatenablog.com/entry/2022/09/03/172957
- https://joker1007.hatenablog.com/entry/2022/09/29/232418
- https://zenn.dev/kawarimidoll/articles/8172a4c29a6653
- https://zenn.dev/yano/articles/neovim_frontend_development_2022
- https://zenn.dev/yutakatay/articles/neovim-plugins-2022

Cursor shape was returned to the default setting.
When I tried the same settings as before, the shape did not change in Replace mode, and when I quit Neovim and returned to terminal, the Neovim settings were taken over, which was annoying.