Make sure to use the language runtime environment installed with mise to add and update libraries
I had not noticed before, but when I ran each setup script the language runtime environment installed by mise was used, but when I ran them as part of a script to set up the development environment for the whole mac, its installed by Homebrew was used.
Maybe it is due to differences in PATH.

Use `mise exec`.
- https://mise.jdx.dev/dev-tools/#mise-exec-mise-x