Simplify setup scripts by using the asdf-flutter patch version
When set up on macOS using asdf-flutter, a zip file is downloaded.
However, asdf-flutter does not support zip files and an error occurs.
There is the Issue and the Pull Request in the repository, but it has not been merged for more than 2 years.
Maybe there is some reason why some environments reproduce the issue and others do not.
Or maybe the repository is not maintained.

See also:
- https://docs.flutter.dev/get-started/install/macos/mobile-ios#download-then-install-flutter
- https://github.com/asdf-community/asdf-flutter/pull/52
- https://github.com/asdf-community/asdf-flutter/issues/33
- https://github.com/asdf-community/asdf-flutter/pull/36